Implement drag and drop primitives, built around Solid state.
- Add `createDraggable` to easily integrate drag behaviour into
components, with the component maintaining control over how it looks and
behaves.
- Add `createDroppable` to register and manage droppable areas.
- Support conditionally enabling and disabling droppables via a
`disabled` option on `createDroppable`.
- Support dragging representations of a draggable through the use of a
`DragOverlay` (which can be removed from the normal document flow).
- Support using sensors to detect and manage drag starts. - Add
`createPointerSensor` as the default sensor, and wrap for convenience in
a `<DragDropSensors/>` component.
- Support customising collision detection algorithms and provide two
initial ones for common usage (`mostIntersectingLayout` and
`closestLayoutCenter`).
- Support using multiple (or nested) `DragDropContext` for containers
isolated from each other.