Automatically detect use of drag overlay and adjust accordingly.
Requiring explicitly setting draggables to ignore transform when a drag
overlay is present feels like an obscure implementation detail that
consumers should not need to be managing. So instead, register
automatically when a drag overlay is used and avoid applying any
transform when using the draggable directive. Note that consumers can
override this explicitly by manually setting the transform that is still
calculated and made available.