Avoid setting noop transform unnecessarily.
Setting a transform creates a new stacking context for the relevant
element. This then changes its order relative to other elements. This is
a good thing for the active draggable as it typically results in it
being ordered above the other static elements. However, if a transform
(even a noop one) is applied to all drag and drop elements then it
becomes more likely that the active draggable will appear under other
elements (such as droppables). So avoid setting a transform
unnecessarily in order to reduce the likelihood of this occurring and
the consumer having to manage ordering more explicitly.

Note that, by removing transforms when a noop is calculated, there is a
possibility that a consumer set transform is removed from an element
when ideally it wouldn't be. In this case the consumer should manage the
transform themselves rather than rely on the directive helper.