Use transformed droppable layouts for collision detection.
It makes more sense to check collisions based on where the droppables
will perceptually be rather than where they were.

This is particularly useful for sorting items of various sizes -
previously there would be visually odd behaviour as items would collide
based on whatever droppable was originally in a position rather than the
droppable currently appearing in that position (due to transforms).

As part of this change also update the sortable logic to:
* Use the sorted indexes for determining if the draggable has moved
  position relatively from last position.
* Compute active draggable corresponding droppable position based on the
  initial droppable it has displaced, rather than the current active
  droppable. This is because the active droppable will now be the active
  draggable's corresponding droppable post sort (based on collision
  detection of current positions).

Note that this breaks `onDragEnd` logic for sortables as the active
droppable will always be the active draggable's corresponding droppable
post sort. This should be addressed subsequently.