Activate pointer sensor if moved enough whilst attached.
If the pointer moves enough whilst still waiting for activation delay
constraint to resolve, automatically activate the sensor assuming that a
drag is intended (as pointer still attached). This should prevent
dragged items appearing to jump when pointer moved quickly during
activation delay period, but a drag was still initiated.

As part of this, store locally the last recorded draggableId for later
activation.