Fix drag with no movement resulting in an incorrect drop.
This was due to collisions not being detected on drag start if layouts
had not changed (the common case). Collisions now always detected by
`dragStart`.

As part of this, `recomputeLayouts` no longer automatically calls
`detectCollisions` when layouts have changed. Instead, it returns a
boolean indicating whether a layout change detected - enabling the
caller to call `detectCollisions` if desired.

For #12