Defer removal of draggable and droppables.
When a draggable/droppable moves containers it may be recreated
depending on the setup. When this happens and a drag is active (such as
when dragging an item between sortable lists) the recreation causes the
active drag to be lost. This is typically not desireable - instead the
active drag should persist. To handle this, defer the cleanup of state
when removing items and only continue with the removal if the item
hasn't been immediately re-added.