Strip translation transform when computing layout.
If a draggable is active when its layout is recomputed, its currently
applied transform will be evaluated as part of its base layout (by
getBoundingClientRect). This results in the transforms
effectively stacking over time and the item being misplaced.

To prevent this, strip any translation transform that is applied on the
element representing the draggable or droppable.

As part of this change, convert Layout into a class to avoid repeating
definition of getter properties.