Compute and apply correct transform for sortable.
A sortable can be transformed either as the active draggable transform
(when no drag overlay is used) or as a droppable transform caused by the
sorting of the list. Correctly compute the correct transform to use in
each case.

As part of this, always store the computed sort transform against the
droppable entry regardless of whether directive used or not. This
ensures consistency in the data (and helps debuggers visualise the
information accurately).

Use a custom render effect for sortables to ensure the appropriately
chosen transform is set on the element. Without this there is more
potential for confusing edge cases caused by the ordering of the
draggable and droppable render effects.