Adjust draggable transform when layout changes.
When a draggable is recreated during active drag, compute the necessary
adjustment to make to transforms to account for the layout shift. Add
this adjustment as a temporary transformer to apply during
displacements. The result of this is a seamless drag even when the
draggable is effectively recreated during the drag (such as can happen
for kanban board style setups).

Note: the current implementation is not optimised.

This transformer approach could be exposed in future to allow
general transform modifications, such as limiting drags to specific
axis etc.