Simplify typings for state.
Whilst technically correct, the presence of undefined in the typing for
state like `droppables` makes it more awkward for consumers of that
state. This is because they have to account for undefined value even
though it will never actually be present (because setting state to
undefined removes it from the state). So simplify the typing and
override where necessary (such as when removing values).