Refactor collision detection to be more context aware.
Whilst it originally felt better to have collision detection abstracted
into considering just layouts, in practice it limits smarter handling
such as tie breaking on active droppable or types.

So now, pass the active draggable and list of droppables directly to
the collision detection algorithm along with some additional useful
context (such as the active droppable id).

As part of this, refactor the existing algorithms to use the new
approach (and tie-break on active droppable to prevent flipping
situations). Remove the old layout based algorithms as no longer used.

This is a breaking change of the interface for collision detection
algorithms and also in the rename of the builtin algorithms:
* closestLayoutCenter -> closestCenter
* mostIntersectingLayout -> mostIntersecting

For #27