Add style helper that ignores no-op transform.
When using the directive form of draggables/droppables/sortables, the
internal setup conditionally sets the transform only when it is not a
no-op. This is desirable as it prevents issues where the presence of a
transform can affect other intended styles (such as z-index). Make it
easier for those not using the directive form to also have this
behaviour.