Detect collisions on adding items if active drag.
Reintroduce automatic detection of collisions on addition of draggables
or droppables when there is an active draggable present.

This is useful when the addition of an item would change the active
droppable by collision. Without this, the previous active droppable
would persist creating a situation where a fast enough move could cause
the active droppable to not be detected as changed.