Avoid redundant work by using untrack over memo.
In this case, the access of the parent state is not relevant at all to
the signal graph and can be safely excluded from consideration.
This avoids redundant work that is done by the memo in reacting to the
change only to discard it when the result is the same.