Store snapshots for previous items in state.
Refactor so that the previous values for draggable and droppable are
snapshots rather than live updating referencese. This enables easier
comparison between active and previous values, which can be useful for
driving transitions and other animations.

As part of this, add state getters to resolve items directly. Longer
term could remove the helpers 'activeDraggable' etc in favour of
using the new state getters directly.

Note: this is a breaking change as what were previously ids in the state
are now object references. For example, `state.active.draggable` is now
a `Draggable`. The previous id value is still available by adding an
`Id` suffix - for example `state.active.draggableId`.