Support smooth transition on drop in draggable.
Handle deferring updating transform as well as transitioning layout
delta at the draggable item level. When a draggable is dropped, it
may be moved to a new container and thus effectively recreated and
reregistered as a draggable. When this happens the draggable will
snap instantly to its new position (as the new transform will be noop).
To allow smooth transitioning from the previous position to the new,
the previous transform is used for a short period and adjusted to
account for any change in the underlying layout it is transforming. The
result is that a draggable can transition seamlessly to its new
position on drop.

However, because CSS transition will use the initial new layout
as a starting position, it is important to only apply the CSS transition
after the adjusted transform is applied (so that it can be the
starting position). To faciliate this, expose an `isTransitioning`
property on the draggable that consumers can use to enable or
disable CSS transitions. In future, consider exposing a `transition`
property as well as applying it by default in the directive mode as
a sensible default.

Note: the approach here is similar to the handling of an active
draggable being recreated as it moves across containers. However, in
that case it is important that shared state is updated so that
collision detection can be computed correctly and ongoing for the
remainder of the drag. This transition animation is more ephemeral and
usage specific - it may differ for sortables/drag overlays etc.