Support transitioning overlay on drop.
When drag end occurs, store the snapshot of the active draggable and use
for transitioning from last known layout and transform to new one. A
snapshot is required because otherwise the overlay would disappear the
moment the draggable ceases to be active post drop preventing it from
transitioning over time.

As part of this, support adding and removing a class to the reference
draggable node for the duration of the overlay transition.