Avoid collapsing transform when snapshotting.
The overlay requires the modified transform in order to know exactly
where the draggable was on drop. Other onDragEnd handlers likely have
the same need. However, only the base transform was snapshotted when
available, preventing this use case. Avoid this logic of collapsing to
either base or modified transform in snapshot function. Instead, let
each consumer decide. For example, update the addDraggable function
to use the base transform.

Aside: In general this management of a 'base' transform on the transform
feels brittle and messy. It might be better to refactor so that sensor
coordinates are stored as the base transform separately on drag.