Recompute layouts when new sortable index set.
Ensure that the layouts are updated when the sortable context changes
as most likely the nodes and their layouts have changed position too.
This is useful for DragOverlay (for example) to have the correct
positions of nodes to target when animating a transition on drop.
Without this, the layouts could be outdated and the overlay would
transition to the wrong location.

This could have also been done in DragOverlay (as the use case lies
there), or even in dragEnd. But the sortable context felt the more
appropriate place as the cause is specific to sortables.

As part of this, also exclude items pending cleanup from having their
layouts updated. Without this, the transition of items handled in
addDraggable would no longer work as the existing layout and current
layout would be the same.

All of this is starting to feel a bit messy though.