Use source draggable as sensible base size.
Typically, folks expect the overlay to be similar in size to the source
draggable. It can be surprising that an overlay would shrink due to no
longer being affected by the source draggables container layout. To
mitigate this, set a minimum size that equals the source draggable and
ensures an overlay can grow, but not be smaller than the draggable.

For those that truly want a smaller overlay, set DragOverlay
min-width and min-height to appropriate values (such as 0).