Remove now redundant untrack usage.
Solid JS 1.4 no longer eagerly tracks changes to the parent when
accessing a specific property or index. As a result, there is no need
to manually ignore the parent with untrack.

Note; If changes to the parent (e.g. property addition/deletion) is
required, Solid exports a $TRACK symbol for that case.