Refactor core to lean more into reactivity.
Notably, make transforms a reactive computation of an array of
transformers rather than a pre-computed set state. For example, make
active draggable transform react to current sensor position rather than
have sensor set the draggable position directly.

To support this reactivity, introduce a `sensorMove` function in place
of the previous `dragMove` function, add functions to manage
transfomers on draggables and droppables, and remove the now
redundant `displace` function.

Make transformers keyed by id and orderable for ease of use and
predictability.

Transformers also open up opportunity for custom constraints on
transforms, such as limiting drag movement to a particular axis by
adding an appropriate trasnformer in `onDragStart`.

As part of this change, clean up many of the typings and reuse defined
types for consistency. And remove attempts at managing animation in the
core - instead animation will be implemented as a separate concept
similar in principle to the auto-animate library.